requirements.txt
fastapi
uvicorn
numpy
pandas
requests
pillow
deepforest
